+++ CHANGELOG.md
@@ -11,6 +11,10 @@ Upgrading to 3.2.x versions requires that you upgrade to 3.2.0 version first.
 - fix infinite redirect loop on anonymous post page (@balsdorf)
 - add history entry when issue is moved between projects (@balsdorf)
 - generate proper message-id in `Issue::close()` (@glensc, 8a046ea)
+- add example slack error logger (@glensc, e7600031f)
+- Fix bug auto switching project on update issues page (@balsdorf, 6ffecfc)
+- Catch exception if invalid message is added to mail queue (@balsdorf, 0e55ae2)
+- add real 'Status Change Date' column and rename old column to 'Status Action Date' (@balsdorf, #277)
 
 [3.2.2]: https://github.com/eventum/eventum/compare/v3.2.1...master